tests/wpt/tests/css/css-toggle/parsing/toggle-trigger-invalid.tentative.html

<!DOCTYPE HTML>
<meta charset="UTF-8">
<title>CSS Toggles: parsing toggle-trigger with invalid values</title>
<link rel="author" title="L. David Baron" href="https://dbaron.org/">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://tabatkins.github.io/css-toggle/#toggle-trigger-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<body>
<script>

test_invalid_value('toggle-trigger', 'none, none');
test_invalid_value('toggle-trigger', 'none set 1');
test_invalid_value('toggle-trigger', 'None set 1');
test_invalid_value('toggle-trigger', 'NONE set 1');
test_invalid_value('toggle-trigger', 'mytoggle set 1 1');
test_invalid_value('toggle-trigger', 'mytoggle set -1');
test_invalid_value('toggle-trigger', 'none, mytoggle');
test_invalid_value('toggle-trigger', 'mytoggle, none');
test_invalid_value('toggle-trigger', 'mytoggle self 1');
test_invalid_value('toggle-trigger', 'mytoggle 1 self');
test_invalid_value('toggle-trigger', 'mytoggle sticky 1');
test_invalid_value('toggle-trigger', 'mytoggle 1 sticky');
test_invalid_value('toggle-trigger', 'mytoggle self set 1');
test_invalid_value('toggle-trigger', 'mytoggle set 1 self');
test_invalid_value('toggle-trigger', 'mytoggle sticky set 1');
test_invalid_value('toggle-trigger', 'mytoggle set 1 sticky');

</script>
